Handling clicks on matches, but it's not pretty.
