+++ feature/eventdetail/src/commonMain/kotlin/com/adammcneilly/pocketleague/feature/eventdetail/EventDetailScreen.kt
@@ -3,18 +3,13 @@ package com.adammcneilly.pocketleague.feature.eventdetail
 import com.adammcneilly.pocketleague.core.displaymodels.EventDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.feature.CommonParcelize
-import com.adammcneilly.pocketleague.data.event.EventRepository
-import com.adammcneilly.pocketleague.data.match.MatchRepository
+import com.adammcneilly.pocketleague.core.models.Match
 import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
-import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
-import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuit.runtime.ui.Ui
 import com.slack.circuit.runtime.ui.ui
-import org.koin.core.component.KoinComponent
-import org.koin.core.component.inject
 
 /**
  * The screen that shows detailed information about the event with the supplied [eventId].
@@ -46,6 +41,10 @@ data class EventDetailScreen(
         data class StageSelected(
             val stageIndex: Int,
         ) : Event
+
+        data class MatchClicked(
+            val matchId: Match.Id,
+        ) : Event
     }
 
     /**
@@ -61,28 +60,6 @@ data class EventDetailScreen(
                             modifier = modifier,
                         )
                     }
-                }
-
-                else -> null
-            }
-        }
-    }
-
-    /**
-     * Factory to create an [EventDetailPresenter].
-     */
-    object PresenterFactory : Presenter.Factory, KoinComponent {
-        private val eventRepository: EventRepository by inject()
-        private val matchRepository: MatchRepository by inject()
-
-        override fun create(screen: Screen, navigator: Navigator, context: CircuitContext): Presenter<*>? {
-            return when (screen) {
-                is EventDetailScreen -> {
-                    EventDetailPresenter(
-                        eventId = com.adammcneilly.pocketleague.core.models.Event.Id(screen.eventId),
-                        eventRepository = eventRepository,
-                        matchRepository = matchRepository,
-                    )
                 }
 
                 else -> null
