+++ shared/app/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/app/PocketLeagueApp.kt
@@ -38,7 +38,7 @@ fun PocketLeagueApp(
                 listOf(
                     FeedScreen.PresenterFactory,
                     MatchDetailScreen.PresenterFactory,
-                    EventDetailScreen.PresenterFactory,
+                    PresenterFactory,
                 ),
             )
             .build()
